move tooltip, loading, and embedded frame to dedicated directories
