+++ components/Workflows/NewSourceFlow/Steps/SourceConnectionForm.js
@@ -1,6 +1,6 @@
 import { useState } from "react"
 
-import EmbeddedFrame from "@components/EmbeddedFrame"
+import EmbeddedFrame from "@components/EmbeddedFrame/EmbeddedFrame"
 import DevelopmentMessage from "@components/Message/DevelopmentMessage"
 import { useSourceConnectLink } from "@hooks/use-source-connect-link"
 import { useSourceFlow } from "@providers/SourceFlowProvider"
